chore: websocket over TLS version
