+++ package.json
@@ -1,8 +1,10 @@
 {
   "dependencies": {
+    "https": "^1.0.0",
     "net": "^1.0.2",
     "openssl-self-signed-certificate": "^1.1.6",
     "tls": "^0.0.1",
-    "url": "^0.11.0"
+    "url": "^0.11.0",
+    "ws": "^7.3.1"
   }
 }
